+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 183:
 
 * Fix a rare case of failed UTF8 validation
 * Verify certificates in client examples
+* Use boost::empty_value
 
 --------------------------------------------------------------------------------